+++ Lib.Models.TinyTransformer/Layers/TransformerBlock.cs
@@ -5,7 +5,7 @@ namespace Lib.Models.TinyTransformer.Layers;
 public class TransformerBlock
 {
     private readonly TinyTransformerConfig _config;
-    public double[] Forward(int[] context)
+    public float[] Forward(int[] context)
     {
         SelfAttentionLayer selfAttentionLayer = new SelfAttentionLayer(_config);
         FeedForwardLayer feedForwardLayer = new FeedForwardLayer(_config);
